@@ -1,6 +1,6 @@
-module github.com/sourabhmandal/gokeycloak/v1
+module github.com/sourabhmandal/gokeycloak
 
-go 1.18
+go 1.20
 
 require (
 	github.com/go-resty/resty/v2 v2.7.0
@@ -9,12 +9,13 @@ require (
 	github.com/pkg/errors v0.9.1
 	github.com/segmentio/ksuid v1.0.4
 	github.com/stretchr/testify v1.8.2
-	golang.org/x/crypto v0.6.0
+	golang.org/x/crypto v0.8.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/net v0.7.0 // indirect
+	github.com/stretchr/objx v0.5.0 // indirect
+	golang.org/x/net v0.9.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
