@@ -1,14 +1,14 @@
-package gocloak
+package gokeycloak
 
 import "context"
 
-func (g *GoCloak) getAttackDetectionURL(realm string, user string, path ...string) string {
+func (g *GoKeycloak) getAttackDetectionURL(realm string, user string, path ...string) string {
 	path = append([]string{g.basePath, g.Config.authAdminRealms, realm, g.Config.attackDetection, user}, path...)
 	return makeURL(path...)
 }
 
 // GetUserBruteForceDetectionStatus fetches a user status regarding brute force protection
-func (g *GoCloak) GetUserBruteForceDetectionStatus(ctx context.Context, accessToken, realm, userID string) (*BruteForceStatus, error) {
+func (g *GoKeycloak) GetUserBruteForceDetectionStatus(ctx context.Context, accessToken, realm, userID string) (*BruteForceStatus, error) {
 	const errMessage = "could not brute force detection Status"
 	var result BruteForceStatus
 
