@@ -1,4 +1,4 @@
-package gocloak
+package gokeycloak
 
 import (
 	"context"
@@ -6,11 +6,11 @@ import (
 
 	"github.com/golang-jwt/jwt/v4"
 	"github.com/segmentio/ksuid"
-	"github.com/sourabhmandal/gokeycloak/v1/pkg/jwx"
+	"github.com/sourabhmandal/gokeycloak/pkg/jwx"
 )
 
 // LoginAdmin performs a login with Admin client
-func (g *GoCloak) LoginAdmin(ctx context.Context, username, password, realm string) (*JWT, error) {
+func (g *GoKeycloak) LoginAdmin(ctx context.Context, username, password, realm string) (*JWT, error) {
 	return g.GetToken(ctx, realm, TokenOptions{
 		ClientID:  StringP(adminClientID),
 		GrantType: StringP("password"),
@@ -20,7 +20,7 @@ func (g *GoCloak) LoginAdmin(ctx context.Context, username, password, realm stri
 }
 
 // LoginClient performs a login with client credentials
-func (g *GoCloak) LoginClient(ctx context.Context, clientID, clientSecret, realm string) (*JWT, error) {
+func (g *GoKeycloak) LoginClient(ctx context.Context, clientID, clientSecret, realm string) (*JWT, error) {
 	return g.GetToken(ctx, realm, TokenOptions{
 		ClientID:     &clientID,
 		ClientSecret: &clientSecret,
@@ -30,7 +30,7 @@ func (g *GoCloak) LoginClient(ctx context.Context, clientID, clientSecret, realm
 
 // LoginClientTokenExchange will exchange the presented token for a user's token
 // Requires Token-Exchange is enabled: https://www.keycloak.org/docs/latest/securing_apps/index.html#_token-exchange
-func (g *GoCloak) LoginClientTokenExchange(ctx context.Context, clientID, token, clientSecret, realm, targetClient, userID string) (*JWT, error) {
+func (g *GoKeycloak) LoginClientTokenExchange(ctx context.Context, clientID, token, clientSecret, realm, targetClient, userID string) (*JWT, error) {
 	tokenOptions := TokenOptions{
 		ClientID:           &clientID,
 		ClientSecret:       &clientSecret,
@@ -46,7 +46,7 @@ func (g *GoCloak) LoginClientTokenExchange(ctx context.Context, clientID, token,
 }
 
 // LoginClientSignedJWT performs a login with client credentials and signed jwt claims
-func (g *GoCloak) LoginClientSignedJWT(
+func (g *GoKeycloak) LoginClientSignedJWT(
 	ctx context.Context,
 	clientID,
 	realm string,
@@ -77,7 +77,7 @@ func (g *GoCloak) LoginClientSignedJWT(
 }
 
 // Login performs a login with user credentials and a client
-func (g *GoCloak) Login(ctx context.Context, clientID, clientSecret, realm, username, password string) (*JWT, error) {
+func (g *GoKeycloak) Login(ctx context.Context, clientID, clientSecret, realm, username, password string) (*JWT, error) {
 	return g.GetToken(ctx, realm, TokenOptions{
 		ClientID:     &clientID,
 		ClientSecret: &clientSecret,
@@ -89,7 +89,7 @@ func (g *GoCloak) Login(ctx context.Context, clientID, clientSecret, realm, user
 }
 
 // LoginOtp performs a login with user credentials and otp token
-func (g *GoCloak) LoginOtp(ctx context.Context, clientID, clientSecret, realm, username, password, totp string) (*JWT, error) {
+func (g *GoKeycloak) LoginOtp(ctx context.Context, clientID, clientSecret, realm, username, password, totp string) (*JWT, error) {
 	return g.GetToken(ctx, realm, TokenOptions{
 		ClientID:     &clientID,
 		ClientSecret: &clientSecret,
@@ -101,7 +101,7 @@ func (g *GoCloak) LoginOtp(ctx context.Context, clientID, clientSecret, realm, u
 }
 
 // GetAuthenticationFlows get all authentication flows from a realm
-func (g *GoCloak) GetAuthenticationFlows(ctx context.Context, token, realm string) ([]*AuthenticationFlowRepresentation, error) {
+func (g *GoKeycloak) GetAuthenticationFlows(ctx context.Context, token, realm string) ([]*AuthenticationFlowRepresentation, error) {
 	const errMessage = "could not retrieve authentication flows"
 	var result []*AuthenticationFlowRepresentation
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -115,7 +115,7 @@ func (g *GoCloak) GetAuthenticationFlows(ctx context.Context, token, realm strin
 }
 
 // GetAuthenticationFlow get an authentication flow with the given ID
-func (g *GoCloak) GetAuthenticationFlow(ctx context.Context, token, realm string, authenticationFlowID string) (*AuthenticationFlowRepresentation, error) {
+func (g *GoKeycloak) GetAuthenticationFlow(ctx context.Context, token, realm string, authenticationFlowID string) (*AuthenticationFlowRepresentation, error) {
 	const errMessage = "could not retrieve authentication flows"
 	var result *AuthenticationFlowRepresentation
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -129,7 +129,7 @@ func (g *GoCloak) GetAuthenticationFlow(ctx context.Context, token, realm string
 }
 
 // CreateAuthenticationFlow creates a new Authentication flow in a realm
-func (g *GoCloak) CreateAuthenticationFlow(ctx context.Context, token, realm string, flow AuthenticationFlowRepresentation) error {
+func (g *GoKeycloak) CreateAuthenticationFlow(ctx context.Context, token, realm string, flow AuthenticationFlowRepresentation) error {
 	const errMessage = "could not create authentication flows"
 	var result []*AuthenticationFlowRepresentation
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -140,7 +140,7 @@ func (g *GoCloak) CreateAuthenticationFlow(ctx context.Context, token, realm str
 }
 
 // UpdateAuthenticationFlow a given Authentication Flow
-func (g *GoCloak) UpdateAuthenticationFlow(ctx context.Context, token, realm string, flow AuthenticationFlowRepresentation, authenticationFlowID string) (*AuthenticationFlowRepresentation, error) {
+func (g *GoKeycloak) UpdateAuthenticationFlow(ctx context.Context, token, realm string, flow AuthenticationFlowRepresentation, authenticationFlowID string) (*AuthenticationFlowRepresentation, error) {
 	const errMessage = "could not create authentication flows"
 	var result *AuthenticationFlowRepresentation
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -154,7 +154,7 @@ func (g *GoCloak) UpdateAuthenticationFlow(ctx context.Context, token, realm str
 }
 
 // DeleteAuthenticationFlow deletes a flow in a realm with the given ID
-func (g *GoCloak) DeleteAuthenticationFlow(ctx context.Context, token, realm, flowID string) error {
+func (g *GoKeycloak) DeleteAuthenticationFlow(ctx context.Context, token, realm, flowID string) error {
 	const errMessage = "could not delete authentication flows"
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
 		Delete(g.getAdminRealmURL(realm, "authentication", "flows", flowID))
@@ -163,7 +163,7 @@ func (g *GoCloak) DeleteAuthenticationFlow(ctx context.Context, token, realm, fl
 }
 
 // GetAuthenticationExecutions retrieves all executions of a given flow
-func (g *GoCloak) GetAuthenticationExecutions(ctx context.Context, token, realm, flow string) ([]*ModifyAuthenticationExecutionRepresentation, error) {
+func (g *GoKeycloak) GetAuthenticationExecutions(ctx context.Context, token, realm, flow string) ([]*ModifyAuthenticationExecutionRepresentation, error) {
 	const errMessage = "could not retrieve authentication flows"
 	var result []*ModifyAuthenticationExecutionRepresentation
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -177,7 +177,7 @@ func (g *GoCloak) GetAuthenticationExecutions(ctx context.Context, token, realm,
 }
 
 // CreateAuthenticationExecution creates a new execution for the given flow name in the given realm
-func (g *GoCloak) CreateAuthenticationExecution(ctx context.Context, token, realm, flow string, execution CreateAuthenticationExecutionRepresentation) error {
+func (g *GoKeycloak) CreateAuthenticationExecution(ctx context.Context, token, realm, flow string, execution CreateAuthenticationExecutionRepresentation) error {
 	const errMessage = "could not create authentication execution"
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).SetBody(execution).
 		Post(g.getAdminRealmURL(realm, "authentication", "flows", flow, "executions", "execution"))
@@ -186,7 +186,7 @@ func (g *GoCloak) CreateAuthenticationExecution(ctx context.Context, token, real
 }
 
 // UpdateAuthenticationExecution updates an authentication execution for the given flow in the given realm
-func (g *GoCloak) UpdateAuthenticationExecution(ctx context.Context, token, realm, flow string, execution ModifyAuthenticationExecutionRepresentation) error {
+func (g *GoKeycloak) UpdateAuthenticationExecution(ctx context.Context, token, realm, flow string, execution ModifyAuthenticationExecutionRepresentation) error {
 	const errMessage = "could not update authentication execution"
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).SetBody(execution).
 		Put(g.getAdminRealmURL(realm, "authentication", "flows", flow, "executions"))
@@ -195,7 +195,7 @@ func (g *GoCloak) UpdateAuthenticationExecution(ctx context.Context, token, real
 }
 
 // DeleteAuthenticationExecution delete a single execution with the given ID
-func (g *GoCloak) DeleteAuthenticationExecution(ctx context.Context, token, realm, executionID string) error {
+func (g *GoKeycloak) DeleteAuthenticationExecution(ctx context.Context, token, realm, executionID string) error {
 	const errMessage = "could not delete authentication execution"
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
 		Delete(g.getAdminRealmURL(realm, "authentication", "executions", executionID))
@@ -204,7 +204,7 @@ func (g *GoCloak) DeleteAuthenticationExecution(ctx context.Context, token, real
 }
 
 // CreateAuthenticationExecutionFlow creates a new execution for the given flow name in the given realm
-func (g *GoCloak) CreateAuthenticationExecutionFlow(ctx context.Context, token, realm, flow string, executionFlow CreateAuthenticationExecutionFlowRepresentation) error {
+func (g *GoKeycloak) CreateAuthenticationExecutionFlow(ctx context.Context, token, realm, flow string, executionFlow CreateAuthenticationExecutionFlowRepresentation) error {
 	const errMessage = "could not create authentication execution flow"
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).SetBody(executionFlow).
 		Post(g.getAdminRealmURL(realm, "authentication", "flows", flow, "executions", "flow"))
@@ -217,7 +217,7 @@ func (g *GoCloak) CreateAuthenticationExecutionFlow(ctx context.Context, token, 
 // ------------------
 
 // CreateIdentityProvider creates an identity provider in a realm
-func (g *GoCloak) CreateIdentityProvider(ctx context.Context, token string, realm string, providerRep IdentityProviderRepresentation) (string, error) {
+func (g *GoKeycloak) CreateIdentityProvider(ctx context.Context, token string, realm string, providerRep IdentityProviderRepresentation) (string, error) {
 	const errMessage = "could not create identity provider"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -232,7 +232,7 @@ func (g *GoCloak) CreateIdentityProvider(ctx context.Context, token string, real
 }
 
 // GetIdentityProviders returns list of identity providers in a realm
-func (g *GoCloak) GetIdentityProviders(ctx context.Context, token, realm string) ([]*IdentityProviderRepresentation, error) {
+func (g *GoKeycloak) GetIdentityProviders(ctx context.Context, token, realm string) ([]*IdentityProviderRepresentation, error) {
 	const errMessage = "could not get identity providers"
 
 	var result []*IdentityProviderRepresentation
@@ -248,7 +248,7 @@ func (g *GoCloak) GetIdentityProviders(ctx context.Context, token, realm string)
 }
 
 // GetIdentityProvider gets the identity provider in a realm
-func (g *GoCloak) GetIdentityProvider(ctx context.Context, token, realm, alias string) (*IdentityProviderRepresentation, error) {
+func (g *GoKeycloak) GetIdentityProvider(ctx context.Context, token, realm, alias string) (*IdentityProviderRepresentation, error) {
 	const errMessage = "could not get identity provider"
 
 	var result IdentityProviderRepresentation
@@ -264,7 +264,7 @@ func (g *GoCloak) GetIdentityProvider(ctx context.Context, token, realm, alias s
 }
 
 // UpdateIdentityProvider updates the identity provider in a realm
-func (g *GoCloak) UpdateIdentityProvider(ctx context.Context, token, realm, alias string, providerRep IdentityProviderRepresentation) error {
+func (g *GoKeycloak) UpdateIdentityProvider(ctx context.Context, token, realm, alias string, providerRep IdentityProviderRepresentation) error {
 	const errMessage = "could not update identity provider"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -275,7 +275,7 @@ func (g *GoCloak) UpdateIdentityProvider(ctx context.Context, token, realm, alia
 }
 
 // DeleteIdentityProvider deletes the identity provider in a realm
-func (g *GoCloak) DeleteIdentityProvider(ctx context.Context, token, realm, alias string) error {
+func (g *GoKeycloak) DeleteIdentityProvider(ctx context.Context, token, realm, alias string) error {
 	const errMessage = "could not delete identity provider"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -285,7 +285,7 @@ func (g *GoCloak) DeleteIdentityProvider(ctx context.Context, token, realm, alia
 }
 
 // ExportIDPPublicBrokerConfig exports the broker config for a given alias
-func (g *GoCloak) ExportIDPPublicBrokerConfig(ctx context.Context, token, realm, alias string) (*string, error) {
+func (g *GoKeycloak) ExportIDPPublicBrokerConfig(ctx context.Context, token, realm, alias string) (*string, error) {
 	const errMessage = "could not get public identity provider configuration"
 
 	resp, err := g.GetRequestWithBearerAuthXMLHeader(ctx, token).
@@ -300,7 +300,7 @@ func (g *GoCloak) ExportIDPPublicBrokerConfig(ctx context.Context, token, realm,
 }
 
 // ImportIdentityProviderConfig parses and returns the identity provider config at a given URL
-func (g *GoCloak) ImportIdentityProviderConfig(ctx context.Context, token, realm, fromURL, providerID string) (map[string]string, error) {
+func (g *GoKeycloak) ImportIdentityProviderConfig(ctx context.Context, token, realm, fromURL, providerID string) (map[string]string, error) {
 	const errMessage = "could not import config"
 
 	result := make(map[string]string)
@@ -320,7 +320,7 @@ func (g *GoCloak) ImportIdentityProviderConfig(ctx context.Context, token, realm
 }
 
 // ImportIdentityProviderConfigFromFile parses and returns the identity provider config from a given file
-func (g *GoCloak) ImportIdentityProviderConfigFromFile(ctx context.Context, token, realm, providerID, fileName string, fileBody io.Reader) (map[string]string, error) {
+func (g *GoKeycloak) ImportIdentityProviderConfigFromFile(ctx context.Context, token, realm, providerID, fileName string, fileBody io.Reader) (map[string]string, error) {
 	const errMessage = "could not import config"
 
 	result := make(map[string]string)
@@ -340,7 +340,7 @@ func (g *GoCloak) ImportIdentityProviderConfigFromFile(ctx context.Context, toke
 }
 
 // CreateIdentityProviderMapper creates an instance of an identity provider mapper associated with the given alias
-func (g *GoCloak) CreateIdentityProviderMapper(ctx context.Context, token, realm, alias string, mapper IdentityProviderMapper) (string, error) {
+func (g *GoKeycloak) CreateIdentityProviderMapper(ctx context.Context, token, realm, alias string, mapper IdentityProviderMapper) (string, error) {
 	const errMessage = "could not create mapper for identity provider"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -355,7 +355,7 @@ func (g *GoCloak) CreateIdentityProviderMapper(ctx context.Context, token, realm
 }
 
 // GetIdentityProviderMapper gets the mapper by id for the given identity provider alias in a realm
-func (g *GoCloak) GetIdentityProviderMapper(ctx context.Context, token string, realm string, alias string, mapperID string) (*IdentityProviderMapper, error) {
+func (g *GoKeycloak) GetIdentityProviderMapper(ctx context.Context, token string, realm string, alias string, mapperID string) (*IdentityProviderMapper, error) {
 	const errMessage = "could not get identity provider mapper"
 
 	result := IdentityProviderMapper{}
@@ -371,7 +371,7 @@ func (g *GoCloak) GetIdentityProviderMapper(ctx context.Context, token string, r
 }
 
 // DeleteIdentityProviderMapper deletes an instance of an identity provider mapper associated with the given alias and mapper ID
-func (g *GoCloak) DeleteIdentityProviderMapper(ctx context.Context, token, realm, alias, mapperID string) error {
+func (g *GoKeycloak) DeleteIdentityProviderMapper(ctx context.Context, token, realm, alias, mapperID string) error {
 	const errMessage = "could not delete mapper for identity provider"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -381,7 +381,7 @@ func (g *GoCloak) DeleteIdentityProviderMapper(ctx context.Context, token, realm
 }
 
 // GetIdentityProviderMappers returns list of mappers associated with an identity provider
-func (g *GoCloak) GetIdentityProviderMappers(ctx context.Context, token, realm, alias string) ([]*IdentityProviderMapper, error) {
+func (g *GoKeycloak) GetIdentityProviderMappers(ctx context.Context, token, realm, alias string) ([]*IdentityProviderMapper, error) {
 	const errMessage = "could not get identity provider mappers"
 
 	var result []*IdentityProviderMapper
@@ -397,7 +397,7 @@ func (g *GoCloak) GetIdentityProviderMappers(ctx context.Context, token, realm, 
 }
 
 // GetIdentityProviderMapperByID gets the mapper of an identity provider
-func (g *GoCloak) GetIdentityProviderMapperByID(ctx context.Context, token, realm, alias, mapperID string) (*IdentityProviderMapper, error) {
+func (g *GoKeycloak) GetIdentityProviderMapperByID(ctx context.Context, token, realm, alias, mapperID string) (*IdentityProviderMapper, error) {
 	const errMessage = "could not get identity provider mappers"
 
 	var result IdentityProviderMapper
@@ -413,7 +413,7 @@ func (g *GoCloak) GetIdentityProviderMapperByID(ctx context.Context, token, real
 }
 
 // UpdateIdentityProviderMapper updates mapper of an identity provider
-func (g *GoCloak) UpdateIdentityProviderMapper(ctx context.Context, token, realm, alias string, mapper IdentityProviderMapper) error {
+func (g *GoKeycloak) UpdateIdentityProviderMapper(ctx context.Context, token, realm, alias string, mapper IdentityProviderMapper) error {
 	const errMessage = "could not update identity provider mapper"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
