@@ -1,5 +1,5 @@
-// Package gocloak is a golang keycloak adaptor.
-package gocloak
+// package gokeycloak is a golang keycloak adaptor.
+package gokeycloak
 
 import (
 	"context"
@@ -15,7 +15,7 @@ import (
 )
 
 // GoCloak provides functionalities to talk to Keycloak.
-type GoCloak struct {
+type GoKeycloak struct {
 	basePath    string
 	certsCache  sync.Map
 	certsLock   sync.Mutex
@@ -39,7 +39,7 @@ func makeURL(path ...string) string {
 }
 
 // GetRequest returns a request for calling endpoints.
-func (g *GoCloak) GetRequest(ctx context.Context) *resty.Request {
+func (g *GoKeycloak) GetRequest(ctx context.Context) *resty.Request {
 	var err HTTPErrorResponse
 	return injectTracingHeaders(
 		ctx, g.restyClient.R().
@@ -49,7 +49,7 @@ func (g *GoCloak) GetRequest(ctx context.Context) *resty.Request {
 }
 
 // GetRequestWithBearerAuthNoCache returns a JSON base request configured with an auth token and no-cache header.
-func (g *GoCloak) GetRequestWithBearerAuthNoCache(ctx context.Context, token string) *resty.Request {
+func (g *GoKeycloak) GetRequestWithBearerAuthNoCache(ctx context.Context, token string) *resty.Request {
 	return g.GetRequest(ctx).
 		SetAuthToken(token).
 		SetHeader("Content-Type", "application/json").
@@ -57,21 +57,21 @@ func (g *GoCloak) GetRequestWithBearerAuthNoCache(ctx context.Context, token str
 }
 
 // GetRequestWithBearerAuth returns a JSON base request configured with an auth token.
-func (g *GoCloak) GetRequestWithBearerAuth(ctx context.Context, token string) *resty.Request {
+func (g *GoKeycloak) GetRequestWithBearerAuth(ctx context.Context, token string) *resty.Request {
 	return g.GetRequest(ctx).
 		SetAuthToken(token).
 		SetHeader("Content-Type", "application/json")
 }
 
 // GetRequestWithBearerAuthXMLHeader returns an XML base request configured with an auth token.
-func (g *GoCloak) GetRequestWithBearerAuthXMLHeader(ctx context.Context, token string) *resty.Request {
+func (g *GoKeycloak) GetRequestWithBearerAuthXMLHeader(ctx context.Context, token string) *resty.Request {
 	return g.GetRequest(ctx).
 		SetAuthToken(token).
 		SetHeader("Content-Type", "application/xml;charset=UTF-8")
 }
 
 // GetRequestWithBasicAuth returns a form data base request configured with basic auth.
-func (g *GoCloak) GetRequestWithBasicAuth(ctx context.Context, clientID, clientSecret string) *resty.Request {
+func (g *GoKeycloak) GetRequestWithBasicAuth(ctx context.Context, clientID, clientSecret string) *resty.Request {
 	req := g.GetRequest(ctx).
 		SetHeader("Content-Type", "application/x-www-form-urlencoded")
 	// Public client doesn't require Basic Auth
@@ -162,8 +162,8 @@ func injectTracingHeaders(ctx context.Context, req *resty.Request) *resty.Reques
 // ===============
 
 // NewClient creates a new Client
-func NewClient(basePath string, options ...func(*GoCloak)) *GoCloak {
-	c := GoCloak{
+func NewClient(basePath string, options ...func(*GoKeycloak)) *GoKeycloak {
+	c := GoKeycloak{
 		basePath:    strings.TrimRight(basePath, urlSeparator),
 		restyClient: resty.New(),
 	}
@@ -183,25 +183,25 @@ func NewClient(basePath string, options ...func(*GoCloak)) *GoCloak {
 
 // RestyClient returns the internal resty g.
 // This can be used to configure the g.
-func (g *GoCloak) RestyClient() *resty.Client {
+func (g *GoKeycloak) RestyClient() *resty.Client {
 	return g.restyClient
 }
 
 // SetRestyClient overwrites the internal resty g.
-func (g *GoCloak) SetRestyClient(restyClient *resty.Client) {
+func (g *GoKeycloak) SetRestyClient(restyClient *resty.Client) {
 	g.restyClient = restyClient
 }
 
 // ==== Functional Options ===
 // SetCertCacheInvalidationTime sets the logout
-func SetCertCacheInvalidationTime(duration time.Duration) func(g *GoCloak) {
-	return func(g *GoCloak) {
+func SetCertCacheInvalidationTime(duration time.Duration) func(g *GoKeycloak) {
+	return func(g *GoKeycloak) {
 		g.Config.CertsInvalidateTime = duration
 	}
 }
 
 // RevokeUserConsents revokes the given user consent.
-func (g *GoCloak) RevokeUserConsents(ctx context.Context, accessToken, realm, userID, clientID string) error {
+func (g *GoKeycloak) RevokeUserConsents(ctx context.Context, accessToken, realm, userID, clientID string) error {
 	const errMessage = "could not revoke consents"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, accessToken).
@@ -211,7 +211,7 @@ func (g *GoCloak) RevokeUserConsents(ctx context.Context, accessToken, realm, us
 }
 
 // LogoutUserSession logs out a single sessions of a user given a session id
-func (g *GoCloak) LogoutUserSession(ctx context.Context, accessToken, realm, session string) error {
+func (g *GoKeycloak) LogoutUserSession(ctx context.Context, accessToken, realm, session string) error {
 	const errMessage = "could not logout"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, accessToken).
@@ -221,7 +221,7 @@ func (g *GoCloak) LogoutUserSession(ctx context.Context, accessToken, realm, ses
 }
 
 // ExecuteActionsEmail executes an actions email
-func (g *GoCloak) ExecuteActionsEmail(ctx context.Context, token, realm string, params ExecuteActionsEmail) error {
+func (g *GoKeycloak) ExecuteActionsEmail(ctx context.Context, token, realm string, params ExecuteActionsEmail) error {
 	const errMessage = "could not execute actions email"
 
 	queryParams, err := GetQueryParams(params)
@@ -238,7 +238,7 @@ func (g *GoCloak) ExecuteActionsEmail(ctx context.Context, token, realm string, 
 }
 
 // CreateComponent creates the given component.
-func (g *GoCloak) CreateComponent(ctx context.Context, token, realm string, component Component) (string, error) {
+func (g *GoKeycloak) CreateComponent(ctx context.Context, token, realm string, component Component) (string, error) {
 	const errMessage = "could not create component"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -253,7 +253,7 @@ func (g *GoCloak) CreateComponent(ctx context.Context, token, realm string, comp
 }
 
 // CreateClient creates the given g.
-func (g *GoCloak) CreateClient(ctx context.Context, accessToken, realm string, newClient Client) (string, error) {
+func (g *GoKeycloak) CreateClient(ctx context.Context, accessToken, realm string, newClient Client) (string, error) {
 	const errMessage = "could not create client"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, accessToken).
@@ -268,7 +268,7 @@ func (g *GoCloak) CreateClient(ctx context.Context, accessToken, realm string, n
 }
 
 // CreateClientRepresentation creates a new client representation
-func (g *GoCloak) CreateClientRepresentation(ctx context.Context, token, realm string, newClient Client) (*Client, error) {
+func (g *GoKeycloak) CreateClientRepresentation(ctx context.Context, token, realm string, newClient Client) (*Client, error) {
 	const errMessage = "could not create client representation"
 
 	var result Client
@@ -286,7 +286,7 @@ func (g *GoCloak) CreateClientRepresentation(ctx context.Context, token, realm s
 }
 
 // CreateClientRole creates a new role for a client
-func (g *GoCloak) CreateClientRole(ctx context.Context, token, realm, idOfClient string, role Role) (string, error) {
+func (g *GoKeycloak) CreateClientRole(ctx context.Context, token, realm, idOfClient string, role Role) (string, error) {
 	const errMessage = "could not create client role"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -301,7 +301,7 @@ func (g *GoCloak) CreateClientRole(ctx context.Context, token, realm, idOfClient
 }
 
 // CreateClientScope creates a new client scope
-func (g *GoCloak) CreateClientScope(ctx context.Context, token, realm string, scope ClientScope) (string, error) {
+func (g *GoKeycloak) CreateClientScope(ctx context.Context, token, realm string, scope ClientScope) (string, error) {
 	const errMessage = "could not create client scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -316,7 +316,7 @@ func (g *GoCloak) CreateClientScope(ctx context.Context, token, realm string, sc
 }
 
 // CreateClientScopeProtocolMapper creates a new protocolMapper under the given client scope
-func (g *GoCloak) CreateClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID string, protocolMapper ProtocolMappers) (string, error) {
+func (g *GoKeycloak) CreateClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID string, protocolMapper ProtocolMappers) (string, error) {
 	const errMessage = "could not create client scope protocol mapper"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -331,7 +331,7 @@ func (g *GoCloak) CreateClientScopeProtocolMapper(ctx context.Context, token, re
 }
 
 // UpdateClient updates the given Client
-func (g *GoCloak) UpdateClient(ctx context.Context, token, realm string, updatedClient Client) error {
+func (g *GoKeycloak) UpdateClient(ctx context.Context, token, realm string, updatedClient Client) error {
 	const errMessage = "could not update client"
 
 	if NilOrEmpty(updatedClient.ID) {
@@ -346,7 +346,7 @@ func (g *GoCloak) UpdateClient(ctx context.Context, token, realm string, updated
 }
 
 // UpdateClientRepresentation updates the given client representation
-func (g *GoCloak) UpdateClientRepresentation(ctx context.Context, accessToken, realm string, updatedClient Client) (*Client, error) {
+func (g *GoKeycloak) UpdateClientRepresentation(ctx context.Context, accessToken, realm string, updatedClient Client) (*Client, error) {
 	const errMessage = "could not update client representation"
 
 	if NilOrEmpty(updatedClient.ID) {
@@ -368,7 +368,7 @@ func (g *GoCloak) UpdateClientRepresentation(ctx context.Context, accessToken, r
 }
 
 // UpdateRole updates the given role.
-func (g *GoCloak) UpdateRole(ctx context.Context, token, realm, idOfClient string, role Role) error {
+func (g *GoKeycloak) UpdateRole(ctx context.Context, token, realm, idOfClient string, role Role) error {
 	const errMessage = "could not update role"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -379,7 +379,7 @@ func (g *GoCloak) UpdateRole(ctx context.Context, token, realm, idOfClient strin
 }
 
 // UpdateClientScope updates the given client scope.
-func (g *GoCloak) UpdateClientScope(ctx context.Context, token, realm string, scope ClientScope) error {
+func (g *GoKeycloak) UpdateClientScope(ctx context.Context, token, realm string, scope ClientScope) error {
 	const errMessage = "could not update client scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -390,7 +390,7 @@ func (g *GoCloak) UpdateClientScope(ctx context.Context, token, realm string, sc
 }
 
 // UpdateClientScopeProtocolMapper updates the given protocol mapper for a client scope
-func (g *GoCloak) UpdateClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID string, protocolMapper ProtocolMappers) error {
+func (g *GoKeycloak) UpdateClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID string, protocolMapper ProtocolMappers) error {
 	const errMessage = "could not update client scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -401,7 +401,7 @@ func (g *GoCloak) UpdateClientScopeProtocolMapper(ctx context.Context, token, re
 }
 
 // DeleteClient deletes a given client
-func (g *GoCloak) DeleteClient(ctx context.Context, token, realm, idOfClient string) error {
+func (g *GoKeycloak) DeleteClient(ctx context.Context, token, realm, idOfClient string) error {
 	const errMessage = "could not delete client"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -411,7 +411,7 @@ func (g *GoCloak) DeleteClient(ctx context.Context, token, realm, idOfClient str
 }
 
 // DeleteComponent deletes the component with the given id.
-func (g *GoCloak) DeleteComponent(ctx context.Context, token, realm, componentID string) error {
+func (g *GoKeycloak) DeleteComponent(ctx context.Context, token, realm, componentID string) error {
 	const errMessage = "could not delete component"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -421,7 +421,7 @@ func (g *GoCloak) DeleteComponent(ctx context.Context, token, realm, componentID
 }
 
 // DeleteClientRepresentation deletes a given client representation.
-func (g *GoCloak) DeleteClientRepresentation(ctx context.Context, accessToken, realm, clientID string) error {
+func (g *GoKeycloak) DeleteClientRepresentation(ctx context.Context, accessToken, realm, clientID string) error {
 	const errMessage = "could not delete client representation"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, accessToken).
@@ -431,7 +431,7 @@ func (g *GoCloak) DeleteClientRepresentation(ctx context.Context, accessToken, r
 }
 
 // DeleteClientRole deletes a given role.
-func (g *GoCloak) DeleteClientRole(ctx context.Context, token, realm, idOfClient, roleName string) error {
+func (g *GoKeycloak) DeleteClientRole(ctx context.Context, token, realm, idOfClient, roleName string) error {
 	const errMessage = "could not delete client role"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -441,7 +441,7 @@ func (g *GoCloak) DeleteClientRole(ctx context.Context, token, realm, idOfClient
 }
 
 // DeleteClientScope deletes the scope with the given id.
-func (g *GoCloak) DeleteClientScope(ctx context.Context, token, realm, scopeID string) error {
+func (g *GoKeycloak) DeleteClientScope(ctx context.Context, token, realm, scopeID string) error {
 	const errMessage = "could not delete client scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -451,7 +451,7 @@ func (g *GoCloak) DeleteClientScope(ctx context.Context, token, realm, scopeID s
 }
 
 // DeleteClientScopeProtocolMapper deletes the given protocol mapper from the client scope
-func (g *GoCloak) DeleteClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID, protocolMapperID string) error {
+func (g *GoKeycloak) DeleteClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID, protocolMapperID string) error {
 	const errMessage = "could not delete client scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -461,7 +461,7 @@ func (g *GoCloak) DeleteClientScopeProtocolMapper(ctx context.Context, token, re
 }
 
 // GetClient returns a client
-func (g *GoCloak) GetClient(ctx context.Context, token, realm, idOfClient string) (*Client, error) {
+func (g *GoKeycloak) GetClient(ctx context.Context, token, realm, idOfClient string) (*Client, error) {
 	const errMessage = "could not get client"
 
 	var result Client
@@ -478,7 +478,7 @@ func (g *GoCloak) GetClient(ctx context.Context, token, realm, idOfClient string
 }
 
 // GetClientRepresentation returns a client representation
-func (g *GoCloak) GetClientRepresentation(ctx context.Context, accessToken, realm, clientID string) (*Client, error) {
+func (g *GoKeycloak) GetClientRepresentation(ctx context.Context, accessToken, realm, clientID string) (*Client, error) {
 	const errMessage = "could not get client representation"
 
 	var result Client
@@ -495,7 +495,7 @@ func (g *GoCloak) GetClientRepresentation(ctx context.Context, accessToken, real
 }
 
 // GetAdapterConfiguration returns a adapter configuration
-func (g *GoCloak) GetAdapterConfiguration(ctx context.Context, accessToken, realm, clientID string) (*AdapterConfiguration, error) {
+func (g *GoKeycloak) GetAdapterConfiguration(ctx context.Context, accessToken, realm, clientID string) (*AdapterConfiguration, error) {
 	const errMessage = "could not get adapter configuration"
 
 	var result AdapterConfiguration
@@ -512,7 +512,7 @@ func (g *GoCloak) GetAdapterConfiguration(ctx context.Context, accessToken, real
 }
 
 // GetClientsDefaultScopes returns a list of the client's default scopes
-func (g *GoCloak) GetClientsDefaultScopes(ctx context.Context, token, realm, idOfClient string) ([]*ClientScope, error) {
+func (g *GoKeycloak) GetClientsDefaultScopes(ctx context.Context, token, realm, idOfClient string) ([]*ClientScope, error) {
 	const errMessage = "could not get clients default scopes"
 
 	var result []*ClientScope
@@ -529,7 +529,7 @@ func (g *GoCloak) GetClientsDefaultScopes(ctx context.Context, token, realm, idO
 }
 
 // AddDefaultScopeToClient adds a client scope to the list of client's default scopes
-func (g *GoCloak) AddDefaultScopeToClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
+func (g *GoKeycloak) AddDefaultScopeToClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
 	const errMessage = "could not add default scope to client"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -539,7 +539,7 @@ func (g *GoCloak) AddDefaultScopeToClient(ctx context.Context, token, realm, idO
 }
 
 // RemoveDefaultScopeFromClient removes a client scope from the list of client's default scopes
-func (g *GoCloak) RemoveDefaultScopeFromClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
+func (g *GoKeycloak) RemoveDefaultScopeFromClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
 	const errMessage = "could not remove default scope from client"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -549,7 +549,7 @@ func (g *GoCloak) RemoveDefaultScopeFromClient(ctx context.Context, token, realm
 }
 
 // GetClientsOptionalScopes returns a list of the client's optional scopes
-func (g *GoCloak) GetClientsOptionalScopes(ctx context.Context, token, realm, idOfClient string) ([]*ClientScope, error) {
+func (g *GoKeycloak) GetClientsOptionalScopes(ctx context.Context, token, realm, idOfClient string) ([]*ClientScope, error) {
 	const errMessage = "could not get clients optional scopes"
 
 	var result []*ClientScope
@@ -566,7 +566,7 @@ func (g *GoCloak) GetClientsOptionalScopes(ctx context.Context, token, realm, id
 }
 
 // AddOptionalScopeToClient adds a client scope to the list of client's optional scopes
-func (g *GoCloak) AddOptionalScopeToClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
+func (g *GoKeycloak) AddOptionalScopeToClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
 	const errMessage = "could not add optional scope to client"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -576,7 +576,7 @@ func (g *GoCloak) AddOptionalScopeToClient(ctx context.Context, token, realm, id
 }
 
 // RemoveOptionalScopeFromClient deletes a client scope from the list of client's optional scopes
-func (g *GoCloak) RemoveOptionalScopeFromClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
+func (g *GoKeycloak) RemoveOptionalScopeFromClient(ctx context.Context, token, realm, idOfClient, scopeID string) error {
 	const errMessage = "could not remove optional scope from client"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -586,7 +586,7 @@ func (g *GoCloak) RemoveOptionalScopeFromClient(ctx context.Context, token, real
 }
 
 // GetDefaultOptionalClientScopes returns a list of default realm optional scopes
-func (g *GoCloak) GetDefaultOptionalClientScopes(ctx context.Context, token, realm string) ([]*ClientScope, error) {
+func (g *GoKeycloak) GetDefaultOptionalClientScopes(ctx context.Context, token, realm string) ([]*ClientScope, error) {
 	const errMessage = "could not get default optional client scopes"
 
 	var result []*ClientScope
@@ -603,7 +603,7 @@ func (g *GoCloak) GetDefaultOptionalClientScopes(ctx context.Context, token, rea
 }
 
 // GetDefaultDefaultClientScopes returns a list of default realm default scopes
-func (g *GoCloak) GetDefaultDefaultClientScopes(ctx context.Context, token, realm string) ([]*ClientScope, error) {
+func (g *GoKeycloak) GetDefaultDefaultClientScopes(ctx context.Context, token, realm string) ([]*ClientScope, error) {
 	const errMessage = "could not get default client scopes"
 
 	var result []*ClientScope
@@ -620,7 +620,7 @@ func (g *GoCloak) GetDefaultDefaultClientScopes(ctx context.Context, token, real
 }
 
 // GetClientScope returns a clientscope
-func (g *GoCloak) GetClientScope(ctx context.Context, token, realm, scopeID string) (*ClientScope, error) {
+func (g *GoKeycloak) GetClientScope(ctx context.Context, token, realm, scopeID string) (*ClientScope, error) {
 	const errMessage = "could not get client scope"
 
 	var result ClientScope
@@ -637,7 +637,7 @@ func (g *GoCloak) GetClientScope(ctx context.Context, token, realm, scopeID stri
 }
 
 // GetClientScopes returns all client scopes
-func (g *GoCloak) GetClientScopes(ctx context.Context, token, realm string) ([]*ClientScope, error) {
+func (g *GoKeycloak) GetClientScopes(ctx context.Context, token, realm string) ([]*ClientScope, error) {
 	const errMessage = "could not get client scopes"
 
 	var result []*ClientScope
@@ -654,7 +654,7 @@ func (g *GoCloak) GetClientScopes(ctx context.Context, token, realm string) ([]*
 }
 
 // GetClientScopeProtocolMappers returns all protocol mappers of a client scope
-func (g *GoCloak) GetClientScopeProtocolMappers(ctx context.Context, token, realm, scopeID string) ([]*ProtocolMappers, error) {
+func (g *GoKeycloak) GetClientScopeProtocolMappers(ctx context.Context, token, realm, scopeID string) ([]*ProtocolMappers, error) {
 	const errMessage = "could not get client scope protocol mappers"
 
 	var result []*ProtocolMappers
@@ -671,7 +671,7 @@ func (g *GoCloak) GetClientScopeProtocolMappers(ctx context.Context, token, real
 }
 
 // GetClientScopeProtocolMapper returns a protocol mapper of a client scope
-func (g *GoCloak) GetClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID, protocolMapperID string) (*ProtocolMappers, error) {
+func (g *GoKeycloak) GetClientScopeProtocolMapper(ctx context.Context, token, realm, scopeID, protocolMapperID string) (*ProtocolMappers, error) {
 	const errMessage = "could not get client scope protocol mappers"
 
 	var result *ProtocolMappers
@@ -688,7 +688,7 @@ func (g *GoCloak) GetClientScopeProtocolMapper(ctx context.Context, token, realm
 }
 
 // GetClientScopeMappings returns all scope mappings for the client
-func (g *GoCloak) GetClientScopeMappings(ctx context.Context, token, realm, idOfClient string) (*MappingsRepresentation, error) {
+func (g *GoKeycloak) GetClientScopeMappings(ctx context.Context, token, realm, idOfClient string) (*MappingsRepresentation, error) {
 	const errMessage = "could not get all scope mappings for the client"
 
 	var result *MappingsRepresentation
@@ -705,7 +705,7 @@ func (g *GoCloak) GetClientScopeMappings(ctx context.Context, token, realm, idOf
 }
 
 // GetClientScopeMappingsRealmRoles returns realm-level roles associated with the client’s scope
-func (g *GoCloak) GetClientScopeMappingsRealmRoles(ctx context.Context, token, realm, idOfClient string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopeMappingsRealmRoles(ctx context.Context, token, realm, idOfClient string) ([]*Role, error) {
 	const errMessage = "could not get realm-level roles with the client’s scope"
 
 	var result []*Role
@@ -722,7 +722,7 @@ func (g *GoCloak) GetClientScopeMappingsRealmRoles(ctx context.Context, token, r
 }
 
 // GetClientScopeMappingsRealmRolesAvailable returns realm-level roles that are available to attach to this client’s scope
-func (g *GoCloak) GetClientScopeMappingsRealmRolesAvailable(ctx context.Context, token, realm, idOfClient string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopeMappingsRealmRolesAvailable(ctx context.Context, token, realm, idOfClient string) ([]*Role, error) {
 	const errMessage = "could not get available realm-level roles with the client’s scope"
 
 	var result []*Role
@@ -739,7 +739,7 @@ func (g *GoCloak) GetClientScopeMappingsRealmRolesAvailable(ctx context.Context,
 }
 
 // CreateClientScopeMappingsRealmRoles create realm-level roles to the client’s scope
-func (g *GoCloak) CreateClientScopeMappingsRealmRoles(ctx context.Context, token, realm, idOfClient string, roles []Role) error {
+func (g *GoKeycloak) CreateClientScopeMappingsRealmRoles(ctx context.Context, token, realm, idOfClient string, roles []Role) error {
 	const errMessage = "could not create realm-level roles to the client’s scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -750,7 +750,7 @@ func (g *GoCloak) CreateClientScopeMappingsRealmRoles(ctx context.Context, token
 }
 
 // DeleteClientScopeMappingsRealmRoles deletes realm-level roles from the client’s scope
-func (g *GoCloak) DeleteClientScopeMappingsRealmRoles(ctx context.Context, token, realm, idOfClient string, roles []Role) error {
+func (g *GoKeycloak) DeleteClientScopeMappingsRealmRoles(ctx context.Context, token, realm, idOfClient string, roles []Role) error {
 	const errMessage = "could not delete realm-level roles from the client’s scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -761,7 +761,7 @@ func (g *GoCloak) DeleteClientScopeMappingsRealmRoles(ctx context.Context, token
 }
 
 // GetClientScopeMappingsClientRoles returns roles associated with a client’s scope
-func (g *GoCloak) GetClientScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string) ([]*Role, error) {
 	const errMessage = "could not get roles associated with a client’s scope"
 
 	var result []*Role
@@ -778,7 +778,7 @@ func (g *GoCloak) GetClientScopeMappingsClientRoles(ctx context.Context, token, 
 }
 
 // GetClientScopeMappingsClientRolesAvailable returns available roles associated with a client’s scope
-func (g *GoCloak) GetClientScopeMappingsClientRolesAvailable(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopeMappingsClientRolesAvailable(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string) ([]*Role, error) {
 	const errMessage = "could not get available roles associated with a client’s scope"
 
 	var result []*Role
@@ -795,7 +795,7 @@ func (g *GoCloak) GetClientScopeMappingsClientRolesAvailable(ctx context.Context
 }
 
 // CreateClientScopeMappingsClientRoles creates client-level roles from the client’s scope
-func (g *GoCloak) CreateClientScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string, roles []Role) error {
+func (g *GoKeycloak) CreateClientScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string, roles []Role) error {
 	const errMessage = "could not create client-level roles from the client’s scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -806,7 +806,7 @@ func (g *GoCloak) CreateClientScopeMappingsClientRoles(ctx context.Context, toke
 }
 
 // DeleteClientScopeMappingsClientRoles deletes client-level roles from the client’s scope
-func (g *GoCloak) DeleteClientScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string, roles []Role) error {
+func (g *GoKeycloak) DeleteClientScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClient, idOfSelectedClient string, roles []Role) error {
 	const errMessage = "could not delete client-level roles from the client’s scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -817,7 +817,7 @@ func (g *GoCloak) DeleteClientScopeMappingsClientRoles(ctx context.Context, toke
 }
 
 // GetClientSecret returns a client's secret
-func (g *GoCloak) GetClientSecret(ctx context.Context, token, realm, idOfClient string) (*CredentialRepresentation, error) {
+func (g *GoKeycloak) GetClientSecret(ctx context.Context, token, realm, idOfClient string) (*CredentialRepresentation, error) {
 	const errMessage = "could not get client secret"
 
 	var result CredentialRepresentation
@@ -834,7 +834,7 @@ func (g *GoCloak) GetClientSecret(ctx context.Context, token, realm, idOfClient 
 }
 
 // GetClientServiceAccount retrieves the service account "user" for a client if enabled
-func (g *GoCloak) GetClientServiceAccount(ctx context.Context, token, realm, idOfClient string) (*User, error) {
+func (g *GoKeycloak) GetClientServiceAccount(ctx context.Context, token, realm, idOfClient string) (*User, error) {
 	const errMessage = "could not get client service account"
 
 	var result User
@@ -850,7 +850,7 @@ func (g *GoCloak) GetClientServiceAccount(ctx context.Context, token, realm, idO
 }
 
 // RegenerateClientSecret triggers the creation of the new client secret.
-func (g *GoCloak) RegenerateClientSecret(ctx context.Context, token, realm, idOfClient string) (*CredentialRepresentation, error) {
+func (g *GoKeycloak) RegenerateClientSecret(ctx context.Context, token, realm, idOfClient string) (*CredentialRepresentation, error) {
 	const errMessage = "could not regenerate client secret"
 
 	var result CredentialRepresentation
@@ -866,7 +866,7 @@ func (g *GoCloak) RegenerateClientSecret(ctx context.Context, token, realm, idOf
 }
 
 // GetClientOfflineSessions returns offline sessions associated with the client
-func (g *GoCloak) GetClientOfflineSessions(ctx context.Context, token, realm, idOfClient string) ([]*UserSessionRepresentation, error) {
+func (g *GoKeycloak) GetClientOfflineSessions(ctx context.Context, token, realm, idOfClient string) ([]*UserSessionRepresentation, error) {
 	const errMessage = "could not get client offline sessions"
 
 	var res []*UserSessionRepresentation
@@ -882,7 +882,7 @@ func (g *GoCloak) GetClientOfflineSessions(ctx context.Context, token, realm, id
 }
 
 // GetClientUserSessions returns user sessions associated with the client
-func (g *GoCloak) GetClientUserSessions(ctx context.Context, token, realm, idOfClient string) ([]*UserSessionRepresentation, error) {
+func (g *GoKeycloak) GetClientUserSessions(ctx context.Context, token, realm, idOfClient string) ([]*UserSessionRepresentation, error) {
 	const errMessage = "could not get client user sessions"
 
 	var res []*UserSessionRepresentation
@@ -898,7 +898,7 @@ func (g *GoCloak) GetClientUserSessions(ctx context.Context, token, realm, idOfC
 }
 
 // CreateClientProtocolMapper creates a protocol mapper in client scope
-func (g *GoCloak) CreateClientProtocolMapper(ctx context.Context, token, realm, idOfClient string, mapper ProtocolMapperRepresentation) (string, error) {
+func (g *GoKeycloak) CreateClientProtocolMapper(ctx context.Context, token, realm, idOfClient string, mapper ProtocolMapperRepresentation) (string, error) {
 	const errMessage = "could not create client protocol mapper"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -913,7 +913,7 @@ func (g *GoCloak) CreateClientProtocolMapper(ctx context.Context, token, realm, 
 }
 
 // UpdateClientProtocolMapper updates a protocol mapper in client scope
-func (g *GoCloak) UpdateClientProtocolMapper(ctx context.Context, token, realm, idOfClient, mapperID string, mapper ProtocolMapperRepresentation) error {
+func (g *GoKeycloak) UpdateClientProtocolMapper(ctx context.Context, token, realm, idOfClient, mapperID string, mapper ProtocolMapperRepresentation) error {
 	const errMessage = "could not update client protocol mapper"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -924,7 +924,7 @@ func (g *GoCloak) UpdateClientProtocolMapper(ctx context.Context, token, realm, 
 }
 
 // DeleteClientProtocolMapper deletes a protocol mapper in client scope
-func (g *GoCloak) DeleteClientProtocolMapper(ctx context.Context, token, realm, idOfClient, mapperID string) error {
+func (g *GoKeycloak) DeleteClientProtocolMapper(ctx context.Context, token, realm, idOfClient, mapperID string) error {
 	const errMessage = "could not delete client protocol mapper"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -934,7 +934,7 @@ func (g *GoCloak) DeleteClientProtocolMapper(ctx context.Context, token, realm, 
 }
 
 // GetKeyStoreConfig get keystoreconfig of the realm
-func (g *GoCloak) GetKeyStoreConfig(ctx context.Context, token, realm string) (*KeyStoreConfig, error) {
+func (g *GoKeycloak) GetKeyStoreConfig(ctx context.Context, token, realm string) (*KeyStoreConfig, error) {
 	const errMessage = "could not get key store config"
 
 	var result KeyStoreConfig
@@ -949,7 +949,7 @@ func (g *GoCloak) GetKeyStoreConfig(ctx context.Context, token, realm string) (*
 	return &result, nil
 }
 
-func (g *GoCloak) getRoleMappings(ctx context.Context, token, realm, path, objectID string) (*MappingsRepresentation, error) {
+func (g *GoKeycloak) getRoleMappings(ctx context.Context, token, realm, path, objectID string) (*MappingsRepresentation, error) {
 	const errMessage = "could not get role mappings"
 
 	var result MappingsRepresentation
@@ -965,17 +965,17 @@ func (g *GoCloak) getRoleMappings(ctx context.Context, token, realm, path, objec
 }
 
 // GetRoleMappingByGroupID gets the role mappings by group
-func (g *GoCloak) GetRoleMappingByGroupID(ctx context.Context, token, realm, groupID string) (*MappingsRepresentation, error) {
+func (g *GoKeycloak) GetRoleMappingByGroupID(ctx context.Context, token, realm, groupID string) (*MappingsRepresentation, error) {
 	return g.getRoleMappings(ctx, token, realm, "groups", groupID)
 }
 
 // GetRoleMappingByUserID gets the role mappings by user
-func (g *GoCloak) GetRoleMappingByUserID(ctx context.Context, token, realm, userID string) (*MappingsRepresentation, error) {
+func (g *GoKeycloak) GetRoleMappingByUserID(ctx context.Context, token, realm, userID string) (*MappingsRepresentation, error) {
 	return g.getRoleMappings(ctx, token, realm, "users", userID)
 }
 
 // GetClientRoles get all roles for the given client in realm
-func (g *GoCloak) GetClientRoles(ctx context.Context, token, realm, idOfClient string, params GetRoleParams) ([]*Role, error) {
+func (g *GoKeycloak) GetClientRoles(ctx context.Context, token, realm, idOfClient string, params GetRoleParams) ([]*Role, error) {
 	const errMessage = "could not get client roles"
 
 	var result []*Role
@@ -997,7 +997,7 @@ func (g *GoCloak) GetClientRoles(ctx context.Context, token, realm, idOfClient s
 }
 
 // GetClientRoleByID gets role for the given client in realm using role ID
-func (g *GoCloak) GetClientRoleByID(ctx context.Context, token, realm, roleID string) (*Role, error) {
+func (g *GoKeycloak) GetClientRoleByID(ctx context.Context, token, realm, roleID string) (*Role, error) {
 	const errMessage = "could not get client role"
 
 	var result Role
@@ -1013,7 +1013,7 @@ func (g *GoCloak) GetClientRoleByID(ctx context.Context, token, realm, roleID st
 }
 
 // GetClientRolesByUserID returns all client roles assigned to the given user
-func (g *GoCloak) GetClientRolesByUserID(ctx context.Context, token, realm, idOfClient, userID string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientRolesByUserID(ctx context.Context, token, realm, idOfClient, userID string) ([]*Role, error) {
 	const errMessage = "could not client roles by user id"
 
 	var result []*Role
@@ -1029,7 +1029,7 @@ func (g *GoCloak) GetClientRolesByUserID(ctx context.Context, token, realm, idOf
 }
 
 // GetClientRolesByGroupID returns all client roles assigned to the given group
-func (g *GoCloak) GetClientRolesByGroupID(ctx context.Context, token, realm, idOfClient, groupID string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientRolesByGroupID(ctx context.Context, token, realm, idOfClient, groupID string) ([]*Role, error) {
 	const errMessage = "could not get client roles by group id"
 
 	var result []*Role
@@ -1045,7 +1045,7 @@ func (g *GoCloak) GetClientRolesByGroupID(ctx context.Context, token, realm, idO
 }
 
 // GetCompositeClientRolesByRoleID returns all client composite roles associated with the given client role
-func (g *GoCloak) GetCompositeClientRolesByRoleID(ctx context.Context, token, realm, idOfClient, roleID string) ([]*Role, error) {
+func (g *GoKeycloak) GetCompositeClientRolesByRoleID(ctx context.Context, token, realm, idOfClient, roleID string) ([]*Role, error) {
 	const errMessage = "could not get composite client roles by role id"
 
 	var result []*Role
@@ -1061,7 +1061,7 @@ func (g *GoCloak) GetCompositeClientRolesByRoleID(ctx context.Context, token, re
 }
 
 // GetCompositeClientRolesByUserID returns all client roles and composite roles assigned to the given user
-func (g *GoCloak) GetCompositeClientRolesByUserID(ctx context.Context, token, realm, idOfClient, userID string) ([]*Role, error) {
+func (g *GoKeycloak) GetCompositeClientRolesByUserID(ctx context.Context, token, realm, idOfClient, userID string) ([]*Role, error) {
 	const errMessage = "could not get composite client roles by user id"
 
 	var result []*Role
@@ -1077,7 +1077,7 @@ func (g *GoCloak) GetCompositeClientRolesByUserID(ctx context.Context, token, re
 }
 
 // GetAvailableClientRolesByUserID returns all available client roles to the given user
-func (g *GoCloak) GetAvailableClientRolesByUserID(ctx context.Context, token, realm, idOfClient, userID string) ([]*Role, error) {
+func (g *GoKeycloak) GetAvailableClientRolesByUserID(ctx context.Context, token, realm, idOfClient, userID string) ([]*Role, error) {
 	const errMessage = "could not get available client roles by user id"
 
 	var result []*Role
@@ -1093,7 +1093,7 @@ func (g *GoCloak) GetAvailableClientRolesByUserID(ctx context.Context, token, re
 }
 
 // GetAvailableClientRolesByGroupID returns all available roles to the given group
-func (g *GoCloak) GetAvailableClientRolesByGroupID(ctx context.Context, token, realm, idOfClient, groupID string) ([]*Role, error) {
+func (g *GoKeycloak) GetAvailableClientRolesByGroupID(ctx context.Context, token, realm, idOfClient, groupID string) ([]*Role, error) {
 	const errMessage = "could not get available client roles by user id"
 
 	var result []*Role
@@ -1109,7 +1109,7 @@ func (g *GoCloak) GetAvailableClientRolesByGroupID(ctx context.Context, token, r
 }
 
 // GetCompositeClientRolesByGroupID returns all client roles and composite roles assigned to the given group
-func (g *GoCloak) GetCompositeClientRolesByGroupID(ctx context.Context, token, realm, idOfClient, groupID string) ([]*Role, error) {
+func (g *GoKeycloak) GetCompositeClientRolesByGroupID(ctx context.Context, token, realm, idOfClient, groupID string) ([]*Role, error) {
 	const errMessage = "could not get composite client roles by group id"
 
 	var result []*Role
@@ -1125,7 +1125,7 @@ func (g *GoCloak) GetCompositeClientRolesByGroupID(ctx context.Context, token, r
 }
 
 // GetClientRole get a role for the given client in a realm by role name
-func (g *GoCloak) GetClientRole(ctx context.Context, token, realm, idOfClient, roleName string) (*Role, error) {
+func (g *GoKeycloak) GetClientRole(ctx context.Context, token, realm, idOfClient, roleName string) (*Role, error) {
 	const errMessage = "could not get client role"
 
 	var result Role
@@ -1141,7 +1141,7 @@ func (g *GoCloak) GetClientRole(ctx context.Context, token, realm, idOfClient, r
 }
 
 // GetClients gets all clients in realm
-func (g *GoCloak) GetClients(ctx context.Context, token, realm string, params GetClientsParams) ([]*Client, error) {
+func (g *GoKeycloak) GetClients(ctx context.Context, token, realm string, params GetClientsParams) ([]*Client, error) {
 	const errMessage = "could not get clients"
 
 	var result []*Client
@@ -1172,7 +1172,7 @@ func UserAttributeContains(attributes map[string][]string, attribute, value stri
 }
 
 // ClearUserCache clears realm cache
-func (g *GoCloak) ClearUserCache(ctx context.Context, token, realm string) error {
+func (g *GoKeycloak) ClearUserCache(ctx context.Context, token, realm string) error {
 	const errMessage = "could not clear user cache"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -1182,7 +1182,7 @@ func (g *GoCloak) ClearUserCache(ctx context.Context, token, realm string) error
 }
 
 // ClearKeysCache clears realm cache
-func (g *GoCloak) ClearKeysCache(ctx context.Context, token, realm string) error {
+func (g *GoKeycloak) ClearKeysCache(ctx context.Context, token, realm string) error {
 	const errMessage = "could not clear keys cache"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -1192,7 +1192,7 @@ func (g *GoCloak) ClearKeysCache(ctx context.Context, token, realm string) error
 }
 
 // AddClientRoleComposite adds roles as composite
-func (g *GoCloak) AddClientRoleComposite(ctx context.Context, token, realm, roleID string, roles []Role) error {
+func (g *GoKeycloak) AddClientRoleComposite(ctx context.Context, token, realm, roleID string, roles []Role) error {
 	const errMessage = "could not add client role composite"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -1203,7 +1203,7 @@ func (g *GoCloak) AddClientRoleComposite(ctx context.Context, token, realm, role
 }
 
 // DeleteClientRoleComposite deletes composites from a role
-func (g *GoCloak) DeleteClientRoleComposite(ctx context.Context, token, realm, roleID string, roles []Role) error {
+func (g *GoKeycloak) DeleteClientRoleComposite(ctx context.Context, token, realm, roleID string, roles []Role) error {
 	const errMessage = "could not delete client role composite"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -1214,7 +1214,7 @@ func (g *GoCloak) DeleteClientRoleComposite(ctx context.Context, token, realm, r
 }
 
 // GetClientScopesScopeMappingsRealmRolesAvailable returns realm-level roles that are available to attach to this client scope
-func (g *GoCloak) GetClientScopesScopeMappingsRealmRolesAvailable(ctx context.Context, token, realm, clientScopeID string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopesScopeMappingsRealmRolesAvailable(ctx context.Context, token, realm, clientScopeID string) ([]*Role, error) {
 	const errMessage = "could not get available realm-level roles with the client-scope"
 
 	var result []*Role
@@ -1231,7 +1231,7 @@ func (g *GoCloak) GetClientScopesScopeMappingsRealmRolesAvailable(ctx context.Co
 }
 
 // GetClientScopesScopeMappingsRealmRoles returns roles associated with a client-scope
-func (g *GoCloak) GetClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string) ([]*Role, error) {
 	const errMessage = "could not get realm-level roles with the client-scope"
 
 	var result []*Role
@@ -1248,7 +1248,7 @@ func (g *GoCloak) GetClientScopesScopeMappingsRealmRoles(ctx context.Context, to
 }
 
 // DeleteClientScopesScopeMappingsRealmRoles deletes realm-level roles from the client-scope
-func (g *GoCloak) DeleteClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string, roles []Role) error {
+func (g *GoKeycloak) DeleteClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string, roles []Role) error {
 	const errMessage = "could not delete realm-level roles from the client-scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -1259,7 +1259,7 @@ func (g *GoCloak) DeleteClientScopesScopeMappingsRealmRoles(ctx context.Context,
 }
 
 // CreateClientScopesScopeMappingsRealmRoles creates realm-level roles to the client scope
-func (g *GoCloak) CreateClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string, roles []Role) error {
+func (g *GoKeycloak) CreateClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string, roles []Role) error {
 	const errMessage = "could not create realm-level roles to the client-scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -1270,7 +1270,7 @@ func (g *GoCloak) CreateClientScopesScopeMappingsRealmRoles(ctx context.Context,
 }
 
 // RegisterRequiredAction creates a required action for a given realm
-func (g *GoCloak) RegisterRequiredAction(ctx context.Context, token string, realm string, requiredAction RequiredActionProviderRepresentation) error {
+func (g *GoKeycloak) RegisterRequiredAction(ctx context.Context, token string, realm string, requiredAction RequiredActionProviderRepresentation) error {
 	const errMessage = "could not create required action"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
@@ -1285,7 +1285,7 @@ func (g *GoCloak) RegisterRequiredAction(ctx context.Context, token string, real
 }
 
 // GetRequiredActions gets a list of required actions for a given realm
-func (g *GoCloak) GetRequiredActions(ctx context.Context, token string, realm string) ([]*RequiredActionProviderRepresentation, error) {
+func (g *GoKeycloak) GetRequiredActions(ctx context.Context, token string, realm string) ([]*RequiredActionProviderRepresentation, error) {
 	const errMessage = "could not get required actions"
 	var result []*RequiredActionProviderRepresentation
 
@@ -1301,7 +1301,7 @@ func (g *GoCloak) GetRequiredActions(ctx context.Context, token string, realm st
 }
 
 // GetRequiredAction gets a required action for a given realm
-func (g *GoCloak) GetRequiredAction(ctx context.Context, token string, realm string, alias string) (*RequiredActionProviderRepresentation, error) {
+func (g *GoKeycloak) GetRequiredAction(ctx context.Context, token string, realm string, alias string) (*RequiredActionProviderRepresentation, error) {
 	const errMessage = "could not get required action"
 	var result RequiredActionProviderRepresentation
 
@@ -1321,7 +1321,7 @@ func (g *GoCloak) GetRequiredAction(ctx context.Context, token string, realm str
 }
 
 // UpdateRequiredAction updates a required action for a given realm
-func (g *GoCloak) UpdateRequiredAction(ctx context.Context, token string, realm string, requiredAction RequiredActionProviderRepresentation) error {
+func (g *GoKeycloak) UpdateRequiredAction(ctx context.Context, token string, realm string, requiredAction RequiredActionProviderRepresentation) error {
 	const errMessage = "could not update required action"
 
 	if NilOrEmpty(requiredAction.ProviderID) {
@@ -1335,7 +1335,7 @@ func (g *GoCloak) UpdateRequiredAction(ctx context.Context, token string, realm 
 }
 
 // DeleteRequiredAction updates a required action for a given realm
-func (g *GoCloak) DeleteRequiredAction(ctx context.Context, token string, realm string, alias string) error {
+func (g *GoKeycloak) DeleteRequiredAction(ctx context.Context, token string, realm string, alias string) error {
 	const errMessage = "could not delete required action"
 
 	if alias == "" {
@@ -1352,7 +1352,7 @@ func (g *GoCloak) DeleteRequiredAction(ctx context.Context, token string, realm 
 }
 
 // CreateClientScopesScopeMappingsClientRoles attaches a client role to a client scope (not client's scope)
-func (g *GoCloak) CreateClientScopesScopeMappingsClientRoles(
+func (g *GoKeycloak) CreateClientScopesScopeMappingsClientRoles(
 	ctx context.Context, token, realm, idOfClientScope, idOfClient string, roles []Role,
 ) error {
 	const errMessage = "could not create client-level roles to the client-scope"
@@ -1367,7 +1367,7 @@ func (g *GoCloak) CreateClientScopesScopeMappingsClientRoles(
 // GetClientScopesScopeMappingsClientRolesAvailable returns available (i.e. not attached via
 // CreateClientScopesScopeMappingsClientRoles) client roles for a specific client, for a client scope
 // (not client's scope).
-func (g *GoCloak) GetClientScopesScopeMappingsClientRolesAvailable(ctx context.Context, token, realm, idOfClientScope, idOfClient string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopesScopeMappingsClientRolesAvailable(ctx context.Context, token, realm, idOfClientScope, idOfClient string) ([]*Role, error) {
 	const errMessage = "could not get available client-level roles with the client-scope"
 
 	var result []*Role
@@ -1385,7 +1385,7 @@ func (g *GoCloak) GetClientScopesScopeMappingsClientRolesAvailable(ctx context.C
 
 // GetClientScopesScopeMappingsClientRoles returns attached client roles for a specific client, for a client scope
 // (not client's scope).
-func (g *GoCloak) GetClientScopesScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClientScope, idOfClient string) ([]*Role, error) {
+func (g *GoKeycloak) GetClientScopesScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClientScope, idOfClient string) ([]*Role, error) {
 	const errMessage = "could not get client-level roles with the client-scope"
 
 	var result []*Role
@@ -1403,7 +1403,7 @@ func (g *GoCloak) GetClientScopesScopeMappingsClientRoles(ctx context.Context, t
 
 // DeleteClientScopesScopeMappingsClientRoles removes attachment of client roles from a client scope
 // (not client's scope).
-func (g *GoCloak) DeleteClientScopesScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClientScope, idOfClient string, roles []Role) error {
+func (g *GoKeycloak) DeleteClientScopesScopeMappingsClientRoles(ctx context.Context, token, realm, idOfClientScope, idOfClient string, roles []Role) error {
 	const errMessage = "could not delete client-level roles from the client-scope"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, token).
