@@ -1,4 +1,4 @@
-package gocloak
+package gokeycloak
 
 import (
 	"context"
@@ -9,19 +9,19 @@ import (
 )
 
 // SetOpenIDConnectEndpoint sets the logout
-func SetOpenIDConnectEndpoint(url string) func(g *GoCloak) {
-	return func(g *GoCloak) {
+func SetOpenIDConnectEndpoint(url string) func(g *GoKeycloak) {
+	return func(g *GoKeycloak) {
 		g.Config.openIDConnect = url
 	}
 }
 
 // // SetRevokeEndpoint sets the revoke endpoint
-// func (g *GoCloak) GetRevokeEndpoint() string {
+// func (g *GoKeycloak) GetRevokeEndpoint() string {
 // 	return g.Config.openIDConnect + "revoke"
 // }
 
 // // SetLogoutEndpoint sets the logout
-// func (g *GoCloak) GetLogoutEndpoint() string {
+// func (g *GoKeycloak) GetLogoutEndpoint() string {
 // 	return g.Config.openIDConnect + "logout"
 // }
 
@@ -30,7 +30,7 @@ func SetOpenIDConnectEndpoint(url string) func(g *GoCloak) {
 //----------------------------------------------------------------------------------
 
 // GetCerts fetches certificates for the given realm from the public /open-id-connect/certs endpoint
-func (g *GoCloak) GetCerts(ctx context.Context, realm string) (*CertResponse, error) {
+func (g *GoKeycloak) GetCerts(ctx context.Context, realm string) (*CertResponse, error) {
 	const errMessage = "could not get certs"
 
 	if cert, ok := g.certsCache.Load(realm); ok {
@@ -57,7 +57,7 @@ func (g *GoCloak) GetCerts(ctx context.Context, realm string) (*CertResponse, er
 	return cert, nil
 }
 
-func (g *GoCloak) getNewCerts(ctx context.Context, realm string) (*CertResponse, error) {
+func (g *GoKeycloak) getNewCerts(ctx context.Context, realm string) (*CertResponse, error) {
 	const errMessage = "could not get newCerts"
 
 	var result CertResponse
@@ -77,7 +77,7 @@ func (g *GoCloak) getNewCerts(ctx context.Context, realm string) (*CertResponse,
 //----------------------------------------------------------------------------------
 
 // GetUserInfo calls the UserInfo endpoint
-func (g *GoCloak) GetUserInfo(ctx context.Context, accessToken, realm string) (*UserInfo, error) {
+func (g *GoKeycloak) GetUserInfo(ctx context.Context, accessToken, realm string) (*UserInfo, error) {
 	const errMessage = "could not get user info"
 
 	var result UserInfo
@@ -93,7 +93,7 @@ func (g *GoCloak) GetUserInfo(ctx context.Context, accessToken, realm string) (*
 }
 
 // GetRawUserInfo calls the UserInfo endpoint and returns a raw json object
-func (g *GoCloak) GetRawUserInfo(ctx context.Context, accessToken, realm string) (map[string]interface{}, error) {
+func (g *GoKeycloak) GetRawUserInfo(ctx context.Context, accessToken, realm string) (map[string]interface{}, error) {
 	const errMessage = "could not get user info"
 
 	var result map[string]interface{}
@@ -112,7 +112,7 @@ func (g *GoCloak) GetRawUserInfo(ctx context.Context, accessToken, realm string)
 //																TOKEN
 //----------------------------------------------------------------------------------
 // RetrospectToken calls the openid-connect introspect endpoint
-func (g *GoCloak) IntrospectToken(ctx context.Context, accessToken, clientID, clientSecret, realm string) (*IntroSpectTokenResult, error) {
+func (g *GoKeycloak) IntrospectToken(ctx context.Context, accessToken, clientID, clientSecret, realm string) (*IntroSpectTokenResult, error) {
 	const errMessage = "could not introspect requesting party token"
 
 	var result IntroSpectTokenResult
@@ -133,7 +133,7 @@ func (g *GoCloak) IntrospectToken(ctx context.Context, accessToken, clientID, cl
 
 // URL: {{keycloak_url}}/realms/{{realm}}/protocol/openid-connect/token
 // GetToken uses TokenOptions to fetch a token.
-func (g *GoCloak) GetToken(ctx context.Context, realm string, options TokenOptions) (*JWT, error) {
+func (g *GoKeycloak) GetToken(ctx context.Context, realm string, options TokenOptions) (*JWT, error) {
 	const errMessage = "could not get token"
 
 	var token JWT
@@ -157,7 +157,7 @@ func (g *GoCloak) GetToken(ctx context.Context, realm string, options TokenOptio
 }
 
 // RevokeToken revokes the passed token. The token can either be an access or refresh token.
-func (g *GoCloak) RevokeToken(ctx context.Context, realm, clientID, clientSecret, refreshToken string) error {
+func (g *GoKeycloak) RevokeToken(ctx context.Context, realm, clientID, clientSecret, refreshToken string) error {
 	const errMessage = "could not revoke token"
 
 	resp, err := g.GetRequestWithBasicAuth(ctx, clientID, clientSecret).
@@ -172,7 +172,7 @@ func (g *GoCloak) RevokeToken(ctx context.Context, realm, clientID, clientSecret
 }
 
 // Logout logs out users with refresh token
-func (g *GoCloak) Logout(ctx context.Context, clientID, clientSecret, realm, refreshToken string) error {
+func (g *GoKeycloak) Logout(ctx context.Context, clientID, clientSecret, realm, refreshToken string) error {
 	const errMessage = "could not logout"
 
 	resp, err := g.GetRequestWithBasicAuth(ctx, clientID, clientSecret).
@@ -186,7 +186,7 @@ func (g *GoCloak) Logout(ctx context.Context, clientID, clientSecret, realm, ref
 }
 
 // LogoutPublicClient performs a logout using a public client and the accessToken.
-func (g *GoCloak) LogoutPublicClient(ctx context.Context, clientID, realm, accessToken, refreshToken string) error {
+func (g *GoKeycloak) LogoutPublicClient(ctx context.Context, clientID, realm, accessToken, refreshToken string) error {
 	const errMessage = "could not logout public client"
 
 	resp, err := g.GetRequestWithBearerAuth(ctx, accessToken).
@@ -201,7 +201,7 @@ func (g *GoCloak) LogoutPublicClient(ctx context.Context, clientID, realm, acces
 
 // RefreshToken refreshes the given token.
 // May return a *APIError with further details about the issue.
-func (g *GoCloak) RefreshToken(ctx context.Context, refreshToken, clientID, clientSecret, realm string) (*JWT, error) {
+func (g *GoKeycloak) RefreshToken(ctx context.Context, refreshToken, clientID, clientSecret, realm string) (*JWT, error) {
 	return g.GetToken(ctx, realm, TokenOptions{
 		ClientID:     &clientID,
 		ClientSecret: &clientSecret,
