@@ -1,4 +1,4 @@
-package gocloak
+package gokeycloak
 
 import (
 	"context"
@@ -7,11 +7,11 @@ import (
 
 	"github.com/golang-jwt/jwt/v4"
 	"github.com/pkg/errors"
-	"github.com/sourabhmandal/gokeycloak/v1/pkg/jwx"
+	"github.com/sourabhmandal/gokeycloak/pkg/jwx"
 )
 
 // GetIssuer gets the issuer of the given realm
-func (g *GoCloak) GetIssuer(ctx context.Context, realm string) (*IssuerResponse, error) {
+func (g *GoKeycloak) GetIssuer(ctx context.Context, realm string) (*IssuerResponse, error) {
 	const errMessage = "could not get issuer"
 
 	var result IssuerResponse
@@ -26,7 +26,7 @@ func (g *GoCloak) GetIssuer(ctx context.Context, realm string) (*IssuerResponse,
 	return &result, nil
 }
 
-func (g *GoCloak) decodeAccessTokenWithClaims(ctx context.Context, accessToken, realm string, claims jwt.Claims) (*jwt.Token, error) {
+func (g *GoKeycloak) decodeAccessTokenWithClaims(ctx context.Context, accessToken, realm string, claims jwt.Claims) (*jwt.Token, error) {
 	const errMessage = "could not decode access token"
 	accessToken = strings.Replace(accessToken, "Bearer ", "", 1)
 
@@ -56,7 +56,7 @@ func (g *GoCloak) decodeAccessTokenWithClaims(ctx context.Context, accessToken, 
 }
 
 // DecodeAccessToken decodes the accessToken
-func (g *GoCloak) DecodeAccessToken(ctx context.Context, accessToken, realm string) (*jwt.Token, *jwt.MapClaims, error) {
+func (g *GoKeycloak) DecodeAccessToken(ctx context.Context, accessToken, realm string) (*jwt.Token, *jwt.MapClaims, error) {
 	claims := jwt.MapClaims{}
 	token, err := g.decodeAccessTokenWithClaims(ctx, accessToken, realm, claims)
 	if err != nil {
@@ -66,6 +66,6 @@ func (g *GoCloak) DecodeAccessToken(ctx context.Context, accessToken, realm stri
 }
 
 // DecodeAccessTokenCustomClaims decodes the accessToken and writes claims into the given claims
-func (g *GoCloak) DecodeAccessTokenCustomClaims(ctx context.Context, accessToken, realm string, claims jwt.Claims) (*jwt.Token, error) {
+func (g *GoKeycloak) DecodeAccessTokenCustomClaims(ctx context.Context, accessToken, realm string, claims jwt.Claims) (*jwt.Token, error) {
 	return g.decodeAccessTokenWithClaims(ctx, accessToken, realm, claims)
 }
